event_searcher/requirements.txt
langchain-openai
langchain-community
langchain
langgraph
embedchain
crewai[tools]
python-dotenv
nicegui
aiosqlite